+++ TP/include/movimento.hpp
@@ -9,13 +9,13 @@ class movimento
 public:
      
     char id = 'M';
-    float tempo; 
-    float objeto;
-    float x_atualizado;
-    float y_atualizado;
+    double tempo; 
+    double objeto;
+    double x_atualizado;
+    double y_atualizado;
      
     
-    movimento(char id ,float tempo , float objeto ,  float x_atualizado, float y_atualizado);
+    movimento(char id ,double tempo , double objeto ,  double x_atualizado, double y_atualizado);
 
     void movimentar(listas& minha_lista);
 };
