+++ TP/src/objeto.cpp
@@ -1,6 +1,6 @@
 #include "include/objeto.hpp"
 
-objetos::objetos(char id , float centro_objeto , float x ,  float y , float largura)
+objetos::objetos(char id , double centro_objeto , double x ,  double y , double largura)
 {
     this->id = id;
     this->id_objeto = centro_objeto;
